refactor lib/models
