+++ lib/model/layers.py
@@ -1,12 +1,139 @@
 #!/usr/bin/env python3
-""" Subpixel Upscaler """
+""" Custom Layers for faceswap.py
+    Layers from:
+        the original https://www.reddit.com/r/deepfakes/ code sample + contribs
+        shoanlu GAN: https://github.com/shaoanlu/faceswap-GAN"""
 
 from __future__ import absolute_import
 
 import tensorflow as tf
 import keras.backend as K
-from keras.engine import Layer
+
+from keras.engine import InputSpec, Layer
+from keras.utils import conv_utils
 from keras.utils.generic_utils import get_custom_objects
+from keras import initializers
+
+
+class PixelShuffler(Layer):
+    """ PixelShuffler layer for Keras
+       by t-ae: https://gist.github.com/t-ae/6e1016cc188104d123676ccef3264981 """
+    # pylint: disable=C0103
+    def __init__(self, size=(2, 2), data_format=None, **kwargs):
+        super(PixelShuffler, self).__init__(**kwargs)
+        self.data_format = K.normalize_data_format(data_format)
+        self.size = conv_utils.normalize_tuple(size, 2, 'size')
+
+    def call(self, inputs, **kwargs):
+
+        input_shape = K.int_shape(inputs)
+        if len(input_shape) != 4:
+            raise ValueError('Inputs should have rank ' +
+                             str(4) +
+                             '; Received input shape:', str(input_shape))
+
+        if self.data_format == 'channels_first':
+            batch_size, c, h, w = input_shape
+            if batch_size is None:
+                batch_size = -1
+            rh, rw = self.size
+            oh, ow = h * rh, w * rw
+            oc = c // (rh * rw)
+
+            out = K.reshape(inputs, (batch_size, rh, rw, oc, h, w))
+            out = K.permute_dimensions(out, (0, 3, 4, 1, 5, 2))
+            out = K.reshape(out, (batch_size, oc, oh, ow))
+        elif self.data_format == 'channels_last':
+            batch_size, h, w, c = input_shape
+            if batch_size is None:
+                batch_size = -1
+            rh, rw = self.size
+            oh, ow = h * rh, w * rw
+            oc = c // (rh * rw)
+
+            out = K.reshape(inputs, (batch_size, h, w, rh, rw, oc))
+            out = K.permute_dimensions(out, (0, 1, 3, 2, 4, 5))
+            out = K.reshape(out, (batch_size, oh, ow, oc))
+        return out
+
+    def compute_output_shape(self, input_shape):
+
+        if len(input_shape) != 4:
+            raise ValueError('Inputs should have rank ' +
+                             str(4) +
+                             '; Received input shape:', str(input_shape))
+
+        if self.data_format == 'channels_first':
+            height = None
+            width = None
+            if input_shape[2] is not None:
+                height = input_shape[2] * self.size[0]
+            if input_shape[3] is not None:
+                width = input_shape[3] * self.size[1]
+            channels = input_shape[1] // self.size[0] // self.size[1]
+
+            if channels * self.size[0] * self.size[1] != input_shape[1]:
+                raise ValueError('channels of input and size are incompatible')
+
+            retval = (input_shape[0],
+                      channels,
+                      height,
+                      width)
+        elif self.data_format == 'channels_last':
+            height = None
+            width = None
+            if input_shape[1] is not None:
+                height = input_shape[1] * self.size[0]
+            if input_shape[2] is not None:
+                width = input_shape[2] * self.size[1]
+            channels = input_shape[3] // self.size[0] // self.size[1]
+
+            if channels * self.size[0] * self.size[1] != input_shape[3]:
+                raise ValueError('channels of input and size are incompatible')
+
+            retval = (input_shape[0],
+                      height,
+                      width,
+                      channels)
+        return retval
+
+    def get_config(self):
+        config = {'size': self.size,
+                  'data_format': self.data_format}
+        base_config = super(PixelShuffler, self).get_config()
+
+        return dict(list(base_config.items()) + list(config.items()))
+
+
+class Scale(Layer):
+    """
+    GAN Custom Scal Layer
+    Code borrows from https://github.com/flyyufelix/cnn_finetune
+    """
+    def __init__(self, weights=None, axis=-1, gamma_init='zero', **kwargs):
+        self.axis = axis
+        self.gamma_init = initializers.get(gamma_init)
+        self.initial_weights = weights
+        super(Scale, self).__init__(**kwargs)
+
+    def build(self, input_shape):
+        self.input_spec = [InputSpec(shape=input_shape)]
+
+        # Compatibility with TensorFlow >= 1.0.0
+        self.gamma = K.variable(self.gamma_init((1,)), name='{}_gamma'.format(self.name))
+        self.trainable_weights = [self.gamma]
+
+        if self.initial_weights is not None:
+            self.set_weights(self.initial_weights)
+            del self.initial_weights
+
+    def call(self, x, mask=None):
+        return self.gamma * x
+
+    def get_config(self):
+        config = {"axis": self.axis}
+        base_config = super(Scale, self).get_config()
+        return dict(list(base_config.items()) + list(config.items()))
 
 
 class SubPixelUpscaling(Layer):
