+++ plugins/train/model/iae.py
@@ -5,7 +5,7 @@ from keras.layers import Concatenate, Dense, Flatten, Input, Reshape
 from keras.layers.convolutional import Conv2D
 from keras.models import Model as KerasModel
 
-from lib.train.nn_blocks import conv, upscale
+from lib.model.nn_blocks import conv, upscale
 from .original import logger, Model as OriginalModel
